Simplified the time function, but I think it was working already after all, after a little more testing

+++ philo/srcs/time.c
@@ -6,7 +6,7 @@
 /*   By: limartin <limartin@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/09/29 15:08:19 by limartin      #+#    #+#                 */
-/*   Updated: 2021/09/29 15:55:03 by limartin      ########   odam.nl         */
+/*   Updated: 2021/09/29 16:12:33 by limartin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,14 +20,14 @@ int ft_start_clock(t_data *d)
 
 int ft_get_ms(t_data *d)
 {
-	int				s;
-	long int		microS;
-	long int		miniS;
+	int	s;
+	int	microS;
+	int	miniS;
 	struct timeval	now;
 
 	gettimeofday(&now, NULL);
-	s = (int)(now.tv_sec - d->start_time.tv_sec);
-	microS = (long int)(now.tv_usec - d->start_time.tv_usec);
+	s = (now.tv_sec - d->start_time.tv_sec);
+	microS = (now.tv_usec - d->start_time.tv_usec);
 	miniS = s * 1000 + microS / 1000;
 	return ((int)miniS);
 }
